requirements.txt
requests
pandas
odfpy
psycopg2-binary
sqlalchemy